# docs/swagger.yaml
basePath: /swagger/things-service
definitions:
...
  common.Page:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  entity.BatchTagInfo:
    properties:
      rel_id:
        description: 对象id
        type: string
      rel_type:
        description: 对象类型 1:设备,2:网关，3:点位
        type: integer
      tags:
        items:
          $ref: '#/definitions/entity.Tag'
        type: array
    type: object
  entity.DeviceCurrentData:
    properties:
      id:
        type: string
      name:
        type: string
      points:
        items:
          $ref: '#/definitions/entity.PointValue'
        type: array
      tags:
        items:
          type: string
        type: array
    type: object
...
  entity.PointValue:
    properties:
      id:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      ts:
        type: string
      value:
        type: integer
    type: object
...
  entity.Tag:
    properties:
      editable:
        description: 是否可以修改编辑。（设备导入的标签不可编辑）
        type: integer
      key:
        description: 标签名称
        type: string
      value:
        description: 标签值
    type: object
...
  model.Tag:
    properties:
      editable:
        description: 是否可编辑（导入的不可编辑）
        type: integer
      id:
        description: 唯一标识
        type: string
      key:
        description: 名称
        type: string
      rel_id:
        description: 关联id
        type: string
      rel_type:
        description: (1:网关、2:设备、3:点位
        type: integer
      user_id:
        description: 用户id
        type: string
      value:
        description: 值
        type: string
    type: object
...
info:
  contact: {}
  description: things-service API 文档.
  title: things-service RESTful API
  version: "1.0"
paths:
...
  /tag:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: rel_id
        in: query
        name: rel_id
        type: string
      - description: key
        in: query
        name: key
        type: string
      - description: value
        in: query
        name: value
        type: string
      - description: editable
        in: query
        name: editable
        type: string
      - description: rel_type
        in: query
        name: rel_type
        type: string
      - description: user_id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Tag'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - Tag
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Tag'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Tag'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - Tag
...
  /tag/all-key-with-value:
    get:
      description: 查询所有distinct标签key,不同的类型会有不同的标签，需要传设备类型。
      parameters:
      - description: 设备类型,1:device,2:gateway,3:point,4:product 默认1
        in: query
        name: type
        type: integer
      - description: 产品id，默认空，则是全部设备，否则为该产品下的所有设备
        in: query
        name: product_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 查询所有distinct标签key 和 values
      tags:
      - Tag
  /tag/all-value:
    get:
      description: 根据标签key查询所有distinct value
      parameters:
      - description: key
        in: query
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 根据标签key查询所有distinct value
      tags:
      - Tag
...
  /tag/batch-get:
    get:
      description: 批量获取一个对象的tag
      parameters:
      - description: 对象id
        in: query
        name: rel_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.BatchTagInfo'
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 批量获取一个对象的tag
      tags:
      - Tag
  /tag/batch-save:
    post:
      description: 批量保存一个对象的tag
      parameters:
      - description: 批量保存tag
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/entity.BatchTagInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 批量保存一个对象的tag
      tags:
      - Tag
...
  /tag/search-device-by-values:
    get:
      description: 根据标签值 模糊搜索 设备数据
      parameters:
      - description: tags value,空格分隔，模糊搜索
        in: query
        name: values
        type: string
      - description: page
        in: query
        name: _page
        type: integer
      - description: pageSize
        in: query
        name: _page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/entity.DeviceCurrentData'
                        type: array
                    type: object
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 根据标签值 模糊搜索 设备数据
      tags:
      - Tag
  /tag/search-point-names:
    get:
      description: 根据标签查询Point名称
      parameters:
      - description: tags标签，key:value,key:value形式
        in: query
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 根据标签查询Point 名称
      tags:
      - Tag
...
swagger: "2.0"